# Conflicts:
#	src/baldr/graphreader.cc
#	src/baldr/graphtile.cc
#	src/baldr/merge.cc
#	src/baldr/pathlocation.cc
#	src/loki/route_action.cc
#	src/loki/worker.cc
#	src/meili/candidate_search.cc
#	src/meili/map_matcher.cc
#	src/meili/routing.cc
#	src/midgard/aabb2.cc
#	src/midgard/logging.cc
#	src/midgard/point2.cc
#	src/midgard/polyline2.cc
#	src/odin/maneuversbuilder.cc
#	src/skadi/sample.cc
#	src/thor/astar.cc
#	src/thor/costmatrix.cc
#	src/thor/isochrone.cc
#	src/thor/multimodal.cc
#	src/thor/route_matcher.cc
#	src/thor/timedistancematrix.cc
#	src/thor/trace_route_action.cc
#	src/thor/trafficalgorithm.cc
#	src/thor/trippathbuilder.cc
#	src/tyr/matrix_serializer.cc
#	src/tyr/navigator.cc
#	src/tyr/route_serializer.cc
#	test/astar.cc
#	valhalla/baldr/graphid.h
#	valhalla/baldr/graphreader.h
#	valhalla/baldr/json.h
#	valhalla/loki/worker.h
#	valhalla/meili/emission_cost_model.h
#	valhalla/meili/state.h
#	valhalla/midgard/pointll.h
#	valhalla/midgard/sequence.h
#	valhalla/tyr/navigator.h

+++ valhalla/thor/timedep.h
@@ -13,7 +13,7 @@ namespace thor {
  * improve performance.
  */
 class TimeDepForward : public AStarPathAlgorithm {
- public:
+public:
   /**
    * Constructor.
    */
@@ -36,11 +36,12 @@ class TimeDepForward : public AStarPathAlgorithm {
    *          each edge).
    */
   virtual std::vector<PathInfo> GetBestPath(odin::Location& origin,
-          odin::Location& dest, baldr::GraphReader& graphreader,
-          const std::shared_ptr<sif::DynamicCost>* mode_costing,
-          const sif::TravelMode mode);
+                                            odin::Location& dest,
+                                            baldr::GraphReader& graphreader,
+                                            const std::shared_ptr<sif::DynamicCost>* mode_costing,
+                                            const sif::TravelMode mode);
 
- protected:
+protected:
   uint32_t origin_tz_index_;
 
   /**
@@ -54,15 +55,18 @@ class TimeDepForward : public AStarPathAlgorithm {
    * @param  pred_idx     Predecessor index into the EdgeLabel list.
    * @param  from_transition True if this method is called from a transition
    *                         edge.
-   * @param  localtime    Current local time.
+   * @param  localtime    Current local time.  Seconds since epoch
    * @param  dest         Location information of the destination.
    * @param  best_path    Best path found so far. Includes the index into
    *                      EdgeLabels and the cost.
    */
   void ExpandForward(baldr::GraphReader& graphreader,
-                     const baldr::GraphId& node, const sif::EdgeLabel& pred,
-                     const uint32_t pred_idx, const bool from_transition,
-                     uint32_t localtime, const odin::Location& dest,
+                     const baldr::GraphId& node,
+                     const sif::EdgeLabel& pred,
+                     const uint32_t pred_idx,
+                     const bool from_transition,
+                     uint64_t localtime,
+                     const odin::Location& dest,
                      std::pair<int32_t, float>& best_path);
 
   /**
@@ -81,7 +85,7 @@ class TimeDepForward : public AStarPathAlgorithm {
  * performance.
  */
 class TimeDepReverse : public AStarPathAlgorithm {
- public:
+public:
   /**
    * Constructor.
    */
@@ -104,11 +108,12 @@ class TimeDepReverse : public AStarPathAlgorithm {
    *          each edge).
    */
   virtual std::vector<PathInfo> GetBestPath(odin::Location& origin,
-          odin::Location& dest, baldr::GraphReader& graphreader,
-          const std::shared_ptr<sif::DynamicCost>* mode_costing,
-          const sif::TravelMode mode);
+                                            odin::Location& dest,
+                                            baldr::GraphReader& graphreader,
+                                            const std::shared_ptr<sif::DynamicCost>* mode_costing,
+                                            const sif::TravelMode mode);
 
- protected:
+protected:
   uint32_t dest_tz_index_;
 
   // Access mode used by the costing method
@@ -137,17 +142,19 @@ class TimeDepReverse : public AStarPathAlgorithm {
    * @param  opp_pred_edge Opposing predecessor directed edge.
    * @param  from_transition True if this method is called from a transition
    *                         edge.
-   * @param  localtime    Current local time.
+   * @param  localtime    Current local time.  Seconds since epoch
    * @param  dest         Location information of the destination.
    * @param  best_path    Best path found so far. Includes the index into
    *                      EdgeLabels and the cost.
    */
   void ExpandReverse(baldr::GraphReader& graphreader,
-                     const baldr::GraphId& node, const sif::BDEdgeLabel& pred,
+                     const baldr::GraphId& node,
+                     const sif::BDEdgeLabel& pred,
                      const uint32_t pred_idx,
                      const baldr::DirectedEdge* opp_pred_edge,
                      const bool from_transition,
-                     uint32_t localtime, const odin::Location& dest,
+                     uint64_t localtime,
+                     const odin::Location& dest,
                      std::pair<int32_t, float>& best_path);
 
   /**
@@ -165,8 +172,8 @@ class TimeDepReverse : public AStarPathAlgorithm {
    * @param  origin       Location information of the origin.
    * @param  destination  Location information of the destination.
    */
-  void SetOrigin(baldr::GraphReader& graphreader, odin::Location& origin,
-                 odin::Location& destination);
+  void
+  SetOrigin(baldr::GraphReader& graphreader, odin::Location& origin, odin::Location& destination);
 
   /**
    * The destination of the reverse path is the origin location. Set the
@@ -190,7 +197,7 @@ class TimeDepReverse : public AStarPathAlgorithm {
   std::vector<PathInfo> FormPath(baldr::GraphReader& graphreader, const uint32_t dest);
 };
 
-}
-}
+} // namespace thor
+} // namespace valhalla
 
-#endif  // VALHALLA_THOR_TIMEDEP_H_
+#endif // VALHALLA_THOR_TIMEDEP_H_
